Grouping, float weight

+++ src/load_and_save.py
@@ -13,7 +13,7 @@ def load_networkx_directed_graph(path: str, weighted=False) -> nx.DiGraph:
     if not weighted:
         graph = nx.read_edgelist(path=path, nodetype=int, create_using=nx.DiGraph())
     else:
-        graph = nx.read_edgelist(path=path, nodetype=int, data=(('weight', int),), create_using=nx.DiGraph())
+        graph = nx.read_edgelist(path=path, nodetype=int, data=(('weight', float),), create_using=nx.DiGraph())
     return graph
